refactor: 컨펌 메세지 상수화, 유틸화
